requirements.txt
django==3.2.7
djangorestframework==3.12.4
gunicorn==20.1.0
django-heroku==0.3.1
djongo
dnspython